requirements.txt
python-mpd2==0.5.5
paho-mqtt==1.3.0
